+++ packages/fs-detectors/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/fs-detectors
 
+## 6.12.0
+
+### Minor Changes
+
+- 3992812: Fix service framework detection when entrypoint is in a subdir of service root.
+
+## 6.11.6
+
+### Patch Changes
+
+- Updated dependencies [85c897e]
+  - @vercel/frameworks@3.30.5
+
 ## 6.11.5
 
 ### Patch Changes